refactor(client): merge BuildClient.ts into client.ts
