--- src/eComMerchant/BuildClient.ts
@@ -1,37 +1,0 @@
-import {
-  ClientBuilder,
-
-  // Import middlewares
-  type AuthMiddlewareOptions, // Required for auth
-  type HttpMiddlewareOptions, // Required for sending HTTP requests
-} from '@commercetools/sdk-client-v2'
-
-const projectKey = import.meta.env.VITE_PROJECT_KEY
-const scopes = [import.meta.env.VITE_SCOPE]
-
-// Configure authMiddlewareOptions
-const authMiddlewareOptions: AuthMiddlewareOptions = {
-  host: import.meta.env.VITE_AUTH_HOST,
-  projectKey,
-  credentials: {
-    clientId: import.meta.env.VITE_CLIENT_ID,
-    clientSecret: import.meta.env.VITE_CLIENT_SECRET,
-  },
-  scopes,
-  fetch,
-}
-
-// Configure httpMiddlewareOptions
-const httpMiddlewareOptions: HttpMiddlewareOptions = {
-  host: import.meta.env.VITE_API_HOST,
-  fetch,
-}
-
-// Export the ClientBuilder
-const ctpClient = new ClientBuilder()
-  .withClientCredentialsFlow(authMiddlewareOptions)
-  .withHttpMiddleware(httpMiddlewareOptions)
-  // .withLoggerMiddleware()
-  .build()
-
-export default ctpClient
